Remove pbtools ns from kineticsTools
[Bug 24835]

[git-p4: depot-paths = "//depot/software/bioinformatics/tools/kineticsTools/": change = 132274]
