+++ bin/summarizeModifications.py
