+++ bin/writeSummaryToCmp.py
