+++ kineticsTools/MixtureEstimationMethods.py
