+++ kineticsTools/ModificationDecode.py
