+++ kineticsTools/p4_c2_medaka_forward_model_weights_half.csv
