--- src/testShared.py
@@ -1,75 +1,0 @@
-#################################################################################
-# Copyright (c) 2011-2013, Pacific Biosciences of California, Inc.
-#
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-# * Redistributions of source code must retain the above copyright
-#   notice, this list of conditions and the following disclaimer.
-# * Redistributions in binary form must reproduce the above copyright
-#   notice, this list of conditions and the following disclaimer in the
-#   documentation and/or other materials provided with the distribution.
-# * Neither the name of Pacific Biosciences nor the names of its
-#   contributors may be used to endorse or promote products derived from
-#   this software without specific prior written permission.
-#
-# NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY
-# THIS LICENSE.  THIS SOFTWARE IS PROVIDED BY PACIFIC BIOSCIENCES AND ITS
-# CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-# PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL PACIFIC BIOSCIENCES OR
-# ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
-# BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
-# IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-#################################################################################
-
-from multiprocessing.process import Process, current_process
-from src.pbtools.kineticsTools.sharedArray import SharedArray
-
-# Test script for making sure that shared memory backed numpy arrays work properly
-# FIXME -- migrate to test dir if possible
-
-
-class Sub(Process):
-
-    def __init__(self, sa):
-        Process.__init__(self)
-        self.sa = sa
-        self.arr = sa.getNumpyWrapper()
-
-    def run(self):
-        import time
-
-        print "self.arr[10] = %f, Process = %s" % (self.arr[10], current_process())
-
-        print self.arr.shape
-
-        n = self.arr.shape[0] - 1
-
-        print "self.arr[%d] = %f, Process = %s" % (n, self.arr[n], current_process())
-        time.sleep(10)
-
-
-class Test:
-
-    def __init__(self):
-        self.sa = SharedArray(dtype="f", shape=50000000)
-
-        self.arr = self.sa.getNumpyWrapper()
-        self.arr[:] = 99.0
-        self.arr[10] = 100.0
-
-    def start(self):
-        proc = Sub(self.sa)
-        proc.start()
-        proc.join()
-
-
-if __name__ == "__main__":
-    tester = Test()
-    tester.start()
