+++ test/test.py
@@ -4,9 +4,9 @@ import platform
 import unittest
 from pbcore.io import CmpH5Reader
 from pbcore.deprecated import ReferenceEntry
-from pbtools.kineticsTools.KineticWorker import KineticWorker
-from pbtools.kineticsTools.ipdModel import IpdModel
-from pbtools.kineticsTools.ReferenceUtils import ReferenceUtils
+from kineticsTools.KineticWorker import KineticWorker
+from kineticsTools.ipdModel import IpdModel
+from kineticsTools.ReferenceUtils import ReferenceUtils
 
 
 class TestSetup(unittest.TestCase):
